Laser now drawn behind ship. Admob retries to initialize ads every X (def 10) seconds now. If there are only fading planetoids left in the scene no damage is done anymore, to prevent those last from killing the player (that was just very annoying.) Updated death-particles to fit better with the style of the rest of the game.
Added settingsManager with a reverseFlight bool to see if it would work. It didn't and now the settings aren't used, I left the files in there however if I decide to try it again later or add other settings.

StopGameModeEmmidiatly is now a bool check in the StopGameMode method, default is false.